Update tic tac toe.py
I plan on making a neural network that will learn how to always win or tie in tic tac toe. At the time of uploading this my plan is to give each possible space a value and have it increase when they use that space and win a game, or decrease when they use that space and lose a game, and then have the cpu pick the highest valued space. Ideally after a couple of games, the cpu will learn either how to go for the win, or which spaces are the best.